fix: replace the hyphen with a dot near the numbers, refactor: create separate folders for pages
