+++ src/components/App/App.tsx
@@ -32,10 +32,10 @@ import { ProfileDocs } from '../../pages/ProfilePage/ProfileDocs/ProfileDocs';
 import { ProfilePage } from '../../pages/ProfilePage/ProfilePage';
 import { ProfileProject } from '../../pages/ProfilePage/ProfileProject/ProfileProject';
 import { ProfileSettings } from '../../pages/ProfilePage/ProfileSettings/ProfileSettings';
-import { LoginPage } from '../../pages/auth/LoginPage';
-import { PasswordResetPage } from '../../pages/auth/PasswordResetPage';
-import { PasswordRestorePage } from '../../pages/auth/PasswordRestorePage';
-import { RegisterPage } from '../../pages/auth/RegisterPage';
+import { LoginPage } from '../../pages/auth/LoginPage/LoginPage';
+import { PasswordResetPage } from '../../pages/auth/PasswordResetPage/PasswordResetPage';
+import { PasswordRestorePage } from '../../pages/auth/PasswordRestorePage/PasswordRestorePage';
+import { RegisterPage } from '../../pages/auth/RegisterPage/RegisterPage';
 import Footer from '../Footer/Footer';
 import { Header } from '../Header/Header';
 import { PopupFeedback } from '../PopupFeedback/PopupFeedback';
