--- src/pages/auth/FormRegister/FormRegister.module.scss
@@ -1,419 +1,0 @@
-.form {
-  --text-container-width: 16%;
-  --grid-column-gap: 61px;
-  --input-side-padding: 12.5%;
-  --input-min-height: 200px;
-  --number-font-size: 128px;
-
-  position: relative;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  box-sizing: border-box;
-  width: 100%;
-
-  @media screen and (width <= 1820px) {
-    --text-container-width: 20%;
-  }
-
-  @media screen and (width <= 1280px) {
-    --grid-column-gap: 36px;
-    --input-min-height: 160px;
-    --input-side-padding: 6.7%; // 80px;
-  }
-
-  @media screen and (width <= 768px) {
-    --grid-column-gap: 13px;
-    --input-side-padding: 2.6%; // 20px;
-    --input-min-height: 95px;
-  }
-
-  @media screen and (width <= 690px) {
-    --grid-column-gap: 13px;
-    --input-side-padding: 5%; // 16px;
-    --input-min-height: 60px;
-  }
-}
-
-.fieldset {
-  position: relative;
-  z-index: 5;
-
-  display: flex;
-  flex-direction: column;
-
-  box-sizing: border-box;
-  width: 100%;
-  margin: 0;
-  padding: 0;
-
-  border: none;
-}
-
-.button {
-  position: absolute;
-  z-index: 3;
-  bottom: 300px;
-  left: 50%;
-  transform: translateX(-50%);
-
-  @media screen and (width <= 1280px) {
-    bottom: 210px;
-    font-size: 48px;
-  }
-
-  @media screen and (width <= 768px) {
-    bottom: 180px;
-    max-width: 240px;
-    max-height: 240px;
-    font-size: 40px;
-  }
-}
-
-.inputContainer {
-  position: relative;
-
-  display: grid;
-  grid-template-areas:
-    'textContainer input hint'
-    '. error .';
-  grid-template-columns: var(--text-container-width) auto;
-  column-gap: var(--grid-column-gap);
-
-  box-sizing: border-box;
-  width: 100%;
-  min-height: var(--input-min-height);
-  padding: 0 var(--input-side-padding);
-
-  border: solid var(--dark-blue, #002d63);
-  border-width: 0 0 2px;
-
-  &:first-of-type {
-    border-top-width: 2px;
-  }
-
-  @media screen and (width <= 768px) {
-    border-width: 0 0 1px;
-
-    &:first-of-type {
-      border-top-width: 1px;
-    }
-  }
-}
-
-.input {
-  --input-font-size: 140px;
-  --input-font-size-min: 48px;
-
-  grid-area: input;
-  align-self: stretch;
-
-  box-sizing: border-box;
-  width: 100%;
-  max-width: 100%;
-  margin: 0;
-  padding: 0;
-
-  font-family: Germano-Regular, sans-serif;
-  font-size: var(--input-font-size);
-  font-weight: 400;
-  line-height: 0.85; // 120px;
-  color: var(--dark-blue, #002d63);
-
-  background-color: transparent;
-  border: none;
-  outline: none;
-  caret-color: var(--electric, #0049d8);
-
-  &::placeholder {
-    color: var(--dark-blue-50, #8096b1);
-    text-align: left;
-    letter-spacing: -0.04em;
-
-    @media screen and (width <= 768px) {
-      letter-spacing: normal;
-    }
-  }
-
-  &:-webkit-autofill,
-  &:-webkit-autofill:hover,
-  &:-webkit-autofill:focus {
-    box-shadow: 0 0 0 1000px #fff inset;
-
-    -webkit-text-fill-color: var(--dark-blue, #002d63);
-  }
-
-  @media screen and (width <= 1280px) {
-    --input-font-size: 100px;
-    --input-font-size-min: 40px;
-
-    line-height: 1.2; // 120px;
-  }
-
-  @media screen and (width <= 768px) {
-    --input-font-size: 64px;
-    --input-font-size-min: 32px;
-
-    line-height: 1; // 64px;
-  }
-
-  @media screen and (width <= 690px) {
-    --input-font-size: 32px;
-    --input-font-size-min: 20px;
-  }
-}
-
-.textContainer {
-  display: flex;
-  grid-area: textContainer;
-  flex-direction: column;
-  align-self: stretch;
-  justify-content: center;
-}
-
-.textNumber {
-  width: 250px;
-
-  font-family: Molot-Regular, sans-serif;
-  font-size: var(--number-font-size);
-  font-weight: 400;
-  font-style: normal;
-  line-height: 1; // 128px;
-  color: var(--white, #fff);
-  text-align: left;
-
-  transition: all 0.2s ease;
-
-  -webkit-text-stroke: 3.5px var(--dark-blue);
-
-  &Error {
-    color: var(--error, #bf1700) !important;
-
-    -webkit-text-stroke: unset;
-  }
-
-  @media screen and (width <= 1280px) {
-    --number-font-size: 80px;
-
-    width: 160px;
-
-    -webkit-text-stroke-width: 2px;
-  }
-
-  @media screen and (width <= 768px) {
-    --number-font-size: 66px;
-
-    width: 120px;
-  }
-
-  @media screen and (width <= 690px) {
-    --number-font-size: 30px;
-
-    width: 54px;
-
-    -webkit-text-stroke-width: 1.2px;
-  }
-}
-
-.textClue {
-  box-sizing: border-box;
-  max-height: 1px;
-
-  font-family: Germano-Regular, sans-serif;
-  font-size: 48px;
-  font-weight: 300;
-  line-height: 1; // 32px;
-  color: var(--dark-blue, #002d63);
-
-  opacity: 0;
-
-  transition: all 0.2s ease;
-
-  &Error {
-    color: var(--error, #bf1700);
-  }
-
-  @media screen and (width <= 1280px) {
-    font-size: 40px;
-  }
-
-  @media screen and (width <= 768px) {
-    font-size: 32px;
-  }
-
-  @media screen and (width <= 690px) {
-    font-size: 24px;
-  }
-}
-
-.inputErrorWrap {
-  overflow-x: hidden; // for word-wrap
-  display: none;
-  grid-area: error;
-
-  width: 100%;
-  min-height: max-content;
-  padding-bottom: 10px;
-
-  &Visible {
-    display: block;
-  }
-}
-
-.inputError {
-  width: 100%;
-  max-width: 100%;
-
-  font-family: Germano-Regular, sans-serif;
-  font-size: 48px;
-  font-weight: 400;
-  line-height: 1;
-  color: var(--error, #bf1700);
-  word-wrap: break-word;
-
-  @media screen and (width <= 1280px) {
-    font-size: 40px;
-  }
-
-  @media screen and (width <= 768px) {
-    font-size: 32px;
-  }
-
-  @media screen and (width <= 360px) {
-    font-size: 16px;
-    line-height: 1.12; // 18px
-  }
-}
-
-.hintIcon {
-  grid-area: hint;
-  align-self: center;
-}
-
-.checkboxContainer {
-  display: flex;
-  flex-direction: row;
-  column-gap: 30px;
-
-  box-sizing: border-box;
-  width: 100%;
-  margin: 28px auto 800px;
-  padding: 0;
-
-  @media screen and (width <= 1280px) {
-    margin-top: 24px;
-    margin-bottom: 622px;
-  }
-
-  @media screen and (width <= 768px) {
-    margin-top: 8px;
-    margin-bottom: 528px;
-  }
-
-  @media screen and (width <= 360px) {
-    column-gap: 15px;
-    margin-top: 12px;
-    margin-bottom: 500px;
-  }
-}
-
-.checkbox {
-  display: block;
-  width: 0;
-  height: 0;
-  opacity: 0;
-}
-
-.checkboxLabel {
-  box-sizing: border-box;
-  width: 60px;
-  height: 60px;
-
-  @media screen and (width <= 768px) {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.checkboxText {
-  margin: 0;
-  padding: 0;
-
-  font-family: Germano, sans-serif;
-  font-size: 64px;
-  font-weight: 300;
-  font-style: normal;
-  line-height: 1.3; // magnification to better match the layout
-  color: var(--dark-blue, #002d63);
-
-  @media screen and (width <= 1280px) {
-    font-size: 48px;
-    line-height: 1.2;
-  }
-
-  @media screen and (width <= 768px) {
-    font-size: 32px;
-  }
-
-  @media screen and (width <= 360px) {
-    font-size: 24px;
-    line-height: 1.13;
-  }
-}
-
-.checkboxTextPolicy {
-  cursor: pointer;
-
-  position: relative;
-  z-index: 6;
-
-  text-decoration-line: underline;
-  text-decoration-thickness: 2px;
-  text-underline-offset: 17px;
-
-  transition: opacity 0.2s ease;
-
-  &:hover {
-    opacity: 0.6;
-  }
-
-  @media screen and (width <= 1280px) {
-    text-underline-offset: 13px;
-  }
-
-  @media screen and (width <= 768px) {
-    text-underline-offset: 9px;
-  }
-
-  @media screen and (width <= 360px) {
-    text-decoration-thickness: 1px;
-    text-underline-offset: 6px;
-  }
-}
-
-.input:focus::placeholder {
-  color: transparent;
-}
-
-.input:focus + .textContainer,
-.input:not(:placeholder-shown) + .textContainer {
-  .textNumber {
-    font-size: calc(var(--number-font-size) * 0.625);
-    color: var(--dark-blue, #002d63);
-
-    -webkit-text-stroke: unset;
-  }
-
-  .textClue {
-    max-height: 50px;
-    padding-top: 10px;
-    padding-bottom: 5px;
-    opacity: 1;
-
-    @media screen and (width <= 690px) {
-      padding-top: 5px;
-    }
-  }
-}
