+++ src/pages/auth/LoginPage/LoginPage.tsx
@@ -1,11 +1,11 @@
 import { FC, FormEvent, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { Main } from '../../components/Main/Main';
-import { useFormAndValidation } from '../../hooks/useFormAndValidation';
-import { FormLogin } from './FormLogin/FormLogin';
-import s from './auth.module.scss';
-import { ROUTE_REGISTER } from '../../constants/constants';
-import useAuth from '../../hooks/useAuth';
+import { Main } from '../../../components/Main/Main';
+import { useFormAndValidation } from '../../../hooks/useFormAndValidation';
+import { FormLogin } from './FormLogin';
+import s from '../auth.module.scss';
+import { ROUTE_REGISTER } from '../../../constants/constants';
+import useAuth from '../../../hooks/useAuth';
 
 export const LoginPage: FC = () => {
   const [isLoading, setIsLoading] = useState(false);
